Refactor: consolidate file write event handlers
Default provider is GitHub.
Set env FILE_PROVIDER to `gitlab` to use GitLab.
