+++ src/events/write-file/index.js
@@ -1,16 +1,12 @@
 const arc = require('@architect/functions')
 const logger = require('@architect/shared/logger')
 const fetch = require('node-fetch')
-const { Octokit } = require('@octokit/rest')
-const octokit = new Octokit({
-  auth: process.env.GITHUB_TOKEN,
-  userAgent: 'Vibrancy https://github.com/barryf/vibrancy'
-})
 
-const githubConfig = {
-  owner: 'barryf',
-  repo: 'content',
-  ref: 'master'
+let provider
+if (process.env.FILE_PROVIDER === 'gitlab') {
+  provider = require('./gitlab')
+} else {
+  provider = require('./github')
 }
 
 function formatPostFile (post) {
@@ -26,29 +22,6 @@ async function formatFileFile (url) {
   const response = await fetch(url)
   const file = await response.buffer()
   return Buffer.from(file).toString('base64')
-}
-
-async function writeGitHubFile (path, method, file, entity) {
-  const params = {
-    path,
-    message: `${entity} ${method}d`,
-    content: file,
-    branch: githubConfig.ref,
-    ...githubConfig
-  }
-  // check if file already exists (and we're updating)
-  try {
-    const response = await octokit.repos.getContent({
-      path,
-      ...githubConfig
-    })
-    params.sha = response.data.sha
-  } catch (err) {
-    if (err.status !== 404) {
-      logger.error('Error fetching file from GitHub', JSON.stringify(err, null, 2))
-    }
-  }
-  return await octokit.repos.createOrUpdateFileContents(params)
 }
 
 exports.handler = async function subscribe (event) {
@@ -82,15 +55,15 @@ exports.handler = async function subscribe (event) {
     path = `files/${body.filePath}`
     //
   } else {
-    logger.error('Unknown folder when writing to GitHub', body.folder)
+    logger.error('Unknown folder when writing file', body.folder)
   }
 
   if (file) {
     try {
-      await writeGitHubFile(path, method, file, entity)
-      logger.info(`Wrote ${entity} to GitHub ${path}`)
+      await provider.writeFile(path, method, file, entity)
+      logger.info(`Wrote ${entity} to ${path}`)
     } catch (err) {
-      logger.error('Error writing file to GitHub', JSON.stringify(err, null, 2))
+      logger.error('Error writing file', JSON.stringify(err, null, 2))
     }
   }
 }
